# assiduidade_system_backend/openapi.yaml
openapi: 3.0.3
info:
  title: FaceClock API
  version: 1.0.0
  description: API principal do MVP de registro de ponto com reconhecimento facial.
servers:
  - url: https://api.faceclock.local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
...
    ConsentCreateRequest:
      type: object
      required: [user_id, term_version, consent_hash, accepted_at]
      properties:
        user_id:
          type: string
          format: uuid
        term_version:
          type: string
        consent_hash:
          type: string
        accepted_at:
          type: string
          format: date-time
        legal_basis:
          type: string
          enum: [CONSENT, LEGAL_OBLIGATION, LEGITIMATE_INTEREST]
    ConsentResponse:
      type: object
      required: [id, user_id, term_version, consent_hash, legal_basis, accepted_at, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        term_version:
          type: string
        consent_hash:
          type: string
        legal_basis:
          type: string
          enum: [CONSENT, LEGAL_OBLIGATION, LEGITIMATE_INTEREST]
        accepted_at:
          type: string
          format: date-time
        revoked_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    ClockRecord:
      type: object
      required: [id, user_id, event_type, recorded_at, source, sync_status]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        device_id:
          type: string
          format: uuid
        event_type:
          type: string
          enum: [ENTRY, BREAK_START, BREAK_END, EXIT]
        recorded_at:
          type: string
          format: date-time
        source:
          type: string
          enum: [ONLINE, OFFLINE_SYNC, MANUAL, INTEGRATION]
        sync_status:
          type: string
          enum: [SYNCED, PENDING, FAILED]
        confidence_score:
          type: number
          format: double
        liveness_score:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
...
    PayrollPushRequest:
      type: object
      required: [provider_name, records]
      properties:
        provider_name:
          type: string
        records:
          type: array
          items:
            $ref: '#/components/schemas/ClockRecord'
    PayrollPushResponse:
      type: object
      required: [batch_id, accepted_records, rejected_records, status]
      properties:
        batch_id:
          type: string
          format: uuid
        accepted_records:
          type: integer
        rejected_records:
          type: integer
        status:
          type: string
paths:
...
  /api/v1/consents:
    post:
      tags: [Consents]
      security:
        - bearerAuth: []
      summary: Registra consentimento do usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentCreateRequest'
      responses:
        '201':
          description: Consentimento registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
  /api/v1/consents/users/{user_id}/active:
    get:
      tags: [Consents]
      security:
        - bearerAuth: []
      summary: Consulta consentimento ativo de um usuario
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Consentimento ativo retornado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '404':
          description: Consentimento ativo nao encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/integrations/payroll/push:
    post:
      tags: [Integrations]
      security:
        - bearerAuth: []
      summary: Envia registros consolidados para integracao externa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayrollPushRequest'
      responses:
        '202':
          description: Lote aceito para processamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayrollPushResponse'
...